package_builder/main_app/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-celery-results==2.5.1
celery==5.3.1
redis==5.0.0
psycopg2-binary==2.9.7
django-prometheus==2.3.0
django-oauth-toolkit==3.0.1
python-dotenv==1.0.0
gunicorn==20.1.0
pytest==7.4.0
pytest-django==4.5.2
django-debug-toolbar==4.1.0
django-redis==5.3.1
djangorestframework-simplejwt==5.2.2
